+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.52.0
 
+- Added `Wizaplace\SDK\Currency\CurrencyService::getAll`
 - Added `Wizaplace\SDK\Currency\CurrencyService::getCountries`
 - Added `Wizaplace\SDK\Currency\CurrencyService::addCountry`
 - Added `Wizaplace\SDK\Currency\CurrencyService::removeCountry`